touch

+++ model/model_v2.py
@@ -83,7 +83,7 @@ class NeuralNetworkModel(LightningModule):
 
     def _common_step(self, batch, batch_idx):
         video, input_label = batch['video'], batch['label']
-        input_label = input_label.to(torch.int64)
+        input_label = input_label.to(torch.float32)
         if self.model_type == "binary":
             output_network = self.forward(video)
             loss = self.loss(self.sigmoid(output_network), input_label.view(input_label.size(0), -1))
